extract redirect utils
