+++ packages/next/shared/lib/router/utils/resolve-rewrites.ts
@@ -1,7 +1,7 @@
 import type { ParsedUrlQuery } from 'querystring'
+import type { Rewrite } from '../../../../lib/load-custom-routes'
 import { getPathMatch } from './path-match'
 import { matchHas, prepareDestination } from './prepare-destination'
-import { Rewrite } from '../../../../lib/load-custom-routes'
 import { removeTrailingSlash } from './remove-trailing-slash'
 import { normalizeLocalePath } from '../../i18n/normalize-locale-path'
 import { removeBasePath } from '../../../../client/remove-base-path'
